fix: lint and prettier fixes to bfp-market

+++ markets/bfp-market/contracts/modules/MarginModule.sol
@@ -9,7 +9,6 @@ import {OwnableStorage} from "@synthetixio/core-contracts/contracts/ownership/Ow
 import {SafeCastI256, SafeCastU256, SafeCastU128} from "@synthetixio/core-contracts/contracts/utils/SafeCast.sol";
 import {ERC165Helper} from "@synthetixio/core-contracts/contracts/utils/ERC165Helper.sol";
 import {IMarginModule} from "../interfaces/IMarginModule.sol";
-import {Order} from "../storage/Order.sol";
 import {PerpMarket} from "../storage/PerpMarket.sol";
 import {PerpMarketConfiguration, SYNTHETIX_USD_MARKET_ID} from "../storage/PerpMarketConfiguration.sol";
 import {Position} from "../storage/Position.sol";
@@ -18,6 +17,8 @@ import {ErrorUtil} from "../utils/ErrorUtil.sol";
 import {MathUtil} from "../utils/MathUtil.sol";
 import {FeatureFlag} from "@synthetixio/core-modules/contracts/storage/FeatureFlag.sol";
 import {Flags} from "../utils/Flags.sol";
+
+/* solhint-disable meta-transactions/no-msg-sender */
 
 contract MarginModule is IMarginModule {
     using SafeCastU256 for uint256;
@@ -433,7 +434,7 @@ contract MarginModule is IMarginModule {
         }
         globalMarginConfig.supportedSynthMarketIds = newSupportedSynthMarketIds;
 
-        for (uint i = 0; i < runtime.lengthBefore; ) {
+        for (uint256 i = 0; i < runtime.lengthBefore; ) {
             uint128 synthMarketId = runtime.previousSupportedSynthMarketIds[i];
 
             // Removing a collateral with a non-zero deposit amount is _not_ allowed. To wind down a collateral,
